+++ src/js/db/connection/index.js
@@ -12,3 +12,5 @@ connection.connect(function (err) {
 	if (err) throw err;
 	console.log("connected as id " + connection.threadId);
 });
+
+module.exports = connection;
