+++ src/js/db/employee/index.js
@@ -1,0 +1,2 @@
+var mysql = require("mysql");
+const connection = require("../connection");
